# [wip] Speed up get_chat_msgs (7 times) and get_chatlist (14% faster)