--- a/src/ephemeral.rs
+++ b/src/ephemeral.rs
@@ -851,7 +851,7 @@ mod tests {
     }
 
     async fn check_msg_was_deleted(t: &TestContext, chat: &Chat, msg_id: MsgId) {
-        sql::housekeeping(&t).await.unwrap();
+        sql::housekeeping(t).await.unwrap();
         let chat_items = chat::get_chat_msgs(t, chat.id, 0, None).await.unwrap();
         // Check that the chat is empty except for possibly info messages:
         for item in &chat_items {