--- a/src/sql.rs
+++ b/src/sql.rs
@@ -603,17 +603,26 @@ pub async fn housekeeping(context: &Context) -> Result<()> {
         warn!(context, "Failed to delete expired messages: {}", err);
     }
 
-    if let Err(err) = remove_unused_files(context).await {
-        warn!(
-            context,
-            "Housekeeping: cannot remove unusued files: {}", err
-        );
-    }
-
     if let Err(err) = start_ephemeral_timers(context).await {
         warn!(
             context,
             "Housekeeping: cannot start ephemeral timers: {}", err
+        );
+    }
+
+    context.set_config_or_warn(Config::LastHousekeeping, Some(&time().to_string())).await;
+
+    let last_time = context.get_config_i64(Config::LastHousekeepingFull).await?;
+    let next_time = last_time + (60 * 60 * 24);
+    if next_time <= time() {
+        info!(context, "Quick Housekeeping done.");
+        return Ok(());
+    }
+
+    if let Err(err) = remove_unused_files(context).await {
+        warn!(
+            context,
+            "Housekeeping: cannot remove unusued files: {}", err
         );
     }
 
@@ -640,9 +649,9 @@ pub async fn housekeeping(context: &Context) -> Result<()> {
         warn!(context, "Failed to run incremental vacuum: {}", err);
     }
 
-    context.set_config_or_warn(Config::LastHousekeeping, Some(&time().to_string())).await;
+    context.set_config_or_warn(Config::LastHousekeepingFull, Some(&time().to_string())).await;
 
-    info!(context, "Housekeeping done.");
+    info!(context, "Full Housekeeping done.");
     Ok(())
 }