@link2xt 2022-03-31
The way things are implemented (as a job), housekeeping will not be run once a minute. Scheduled jobs are run only when the loop is interrupted. So there is no guarantee it will be run even once an hour.

Ephemeral task (scheduled by schedule_ephemeral_task) is now broken, it emits event about the change, but when the chat/chatlist is reloaded, there is no change, because housekeeping job will be run at different time.

For these reasons I think actual message deletion should happen in ephemeral task, not in housekeeping.