@hpk42 2022-04-01
> The way things are implemented (as a job), housekeeping will not be run once a minute. Scheduled jobs are run only when the loop is interrupted. So there is no guarantee it will be run even once an hour.
> 
> Ephemeral task (scheduled by schedule_ephemeral_task) is now broken, it emits event about the change, but when the chat/chatlist is reloaded, there is no change, because housekeeping job will be run at different time.
> 
> For these reasons I think actual message deletion should happen in ephemeral task, not in housekeeping.

right, this occured to me overnight as well.  The Job system is bound to the imap-loop pretty much and it doesn't run often enough.  I am not sure what the best approach is to get something that runs every "time_remaining_till_next_deletion".  In any case, this PR can not be merged as is, so marking it as WIP.