@Hocuri 2022-04-01
> I am not sure what the best approach is to get something that runs every "time_remaining_till_next_deletion". In any case, this PR can not be merged as is, so marking it as WIP.

Since async tasks are cheap, just spawn a task for it in `start_io()`? (and probably stop it in `stop_io()` in order not to get problems with imex and so on).