@hpk42 2022-04-07
today i tackled https://github.com/deltachat/deltachat-core-rust/pull/3194 with @hocuri based on this pr so closing here.